# Replace dictionary with two queues for read map