A major test of data structure change: The query-to-subject(s) map. Previously it was a dictionary of sets. Now it is two queues with equal length for queries and subject sets. Notes:

1. An alternative data structure is a queue of tuple of query/subject(s) pairs. Its performance is moderately lower than the current one. However it may be more readable.
1. `deque` is used for the `Plain` mapper, because it is slightly faster than native `list` according to my tests. However `list` may be more readable and makes unit test easier.

Significant performance improvement was observed:

`align.parse_align_file`: 319 ms ± 3.45 ms => 298 ms ± 1.48 ms (~6.5% improvement)
`workflow.classify`: 370 ms ± 4.3 ms => 359 ms ± 3.2 ms